+++ protected/humhub/libs/BaseSettingsManager.php
@@ -8,8 +8,10 @@
 
 namespace humhub\libs;
 
+use humhub\components\InstallationState;
 use humhub\components\SettingActiveRecord;
 use humhub\exceptions\InvalidArgumentTypeException;
+use humhub\models\Setting;
 use Stringable;
 use Yii;
 use yii\base\Component;
@@ -56,8 +58,9 @@ abstract class BaseSettingsManager extends Component
             throw new InvalidConfigException('Module id not set!', 2);
         }
 
-        if (Yii::$app->isDatabaseInstalled()) {
+        try {
             $this->loadValues();
+        } catch (\Exception $e) {
         }
 
         parent::init();
@@ -153,6 +156,8 @@ abstract class BaseSettingsManager extends Component
      */
     public function get(string $name, $default = null)
     {
+        $name = Setting::fixDeprecatedSettingKeys($name);
+
         $value = $this->_loaded[$name] ?? null;
 
         // make sure it is an int, if it is possible
@@ -308,6 +313,6 @@ abstract class BaseSettingsManager extends Component
      */
     public static function isDatabaseInstalled(): bool
     {
-        return Yii::$app->isDatabaseInstalled(true);
+        return Yii::$app->installationState->hasState(InstallationState::STATE_DATABASE_CREATED);
     }
 }