+++ protected/humhub/modules/admin/messages/nn-NO/information.php
@@ -26,6 +26,7 @@ return [
     'Done' => '',
     'Driver' => '',
     'Dynamic Config' => '',
+    'Enable <a href="{url}">Link Redirection Service</a>' => '',
     'Error' => '',
     'Flush entries' => '',
     'HumHub' => '',