+++ protected/humhub/modules/comment/messages/da/notification.php
@@ -1,5 +1,5 @@
 <?php
-return array (
+return [
   'Comments' => 'Kommentarer',
   'Receive Notifications when someone comments on my own or a following post.' => 'Modtag notifikationer når nogen kommenterer på et opslag jeg har lavet, eller som jeg følger.',
   'View Online' => 'Se Online',
@@ -9,9 +9,9 @@ return array (
   '{displayNames} commented {contentTitle}.' => '{displayNames} kommenterede på {contentTitle}',
   '{displayNames} just commented your {contentTitle}' => '{displayNames} kommenterede på dit opslag {contentTitle}',
   '{displayNames} just commented your {contentTitle} in space {space}' => '{displayNames} kommenterede på dit opslag {contentTitle} på siden {space}',
-  '{displayName} commented {contentTitle}' => '',
-  '{displayName} commented {contentTitle} in space {space}' => '',
-  '{displayName} commented {contentTitle}.' => '',
-  '{displayName} just commented your {contentTitle}' => '',
-  '{displayName} just commented your {contentTitle} in space {space}' => '',
-);
+  '{displayName} commented {contentTitle}' => '{displayName} kommenterede {contentTitle}.',
+  '{displayName} commented {contentTitle} in space {space}' => '{displayName} kommenterede {contentTitle}. i space {space}',
+  '{displayName} commented {contentTitle}.' => '{displayName} kommenterede{contentTitle}.',
+  '{displayName} just commented your {contentTitle}' => '{displayName} har lige kommenteret på {contentTitle}',
+  '{displayName} just commented your {contentTitle} in space {space}' => '{displayName} char lige kommenteret på {contentTitle} i  {space}',
+];