+++ protected/humhub/modules/comment/messages/da/notifications.php
@@ -1,5 +1,4 @@
 <?php
-
 return [
-    'Your comment \'{commentText}\' has been deleted by {displayName} for \'{reason}\'' => '',
+  'Your comment \'{commentText}\' has been deleted by {displayName} for \'{reason}\'' => 'Din kommentar \'{commentText}\' er blevet slettet af {displayName} grundet\'{reason}\'.',
 ];