+++ protected/humhub/modules/comment/messages/da/permissions.php
@@ -1,8 +1,7 @@
 <?php
-
 return [
-    'Allows the user to add comments' => '',
-    'Can manage (e.g. archive, stick, move or delete) arbitrary content' => '',
-    'Create comment' => '',
-    'Manage content' => '',
+  'Allows the user to add comments' => 'Giver brugeren mulighed for at tilføje kommentarer',
+  'Can manage (e.g. archive, stick, move or delete) arbitrary content' => 'Kan administrere (f.eks. arkivere, klistre, flytte eller slette) vilkårligt indhold',
+  'Create comment' => 'Opret kommentar',
+  'Manage content' => 'Administrer indhold',
 ];