+++ protected/humhub/modules/file/messages/th/base.php
@@ -22,4 +22,7 @@ return [
     'Attach a video' => '',
     'Attach an audio message' => '',
     'Attach an image' => '',
+    'Share' => '',
+    'Share as Post' => '',
+    'Share in {targetDisplayName}' => '',
 ];